perfetto: build the x64 CPU check for the baseline ISA
When enable_perfetto_x64_cpu_opt is set every file is compiled with
-mavx2/-msse4.2/etc., including CheckCpuOptimizations() itself. The
compiler then emits those instructions inside the check, so on a CPU
that lacks them the binary hits an illegal instruction (SIGILL) before
the check can print its graceful error.
See https://github.com/google/perfetto/discussions/5138.

What this does:

- Moves CheckCpuOptimizations() into its own file,
  check_cpu_optimizations.cc, built in a target that opts out of the
  CPU optimization flags, so the check runs on any x86_64 CPU.
- Pulls those flags out of config("default") into a dedicated
  config("x64_cpu_opt"), added to default_configs so every other target
  still gets them unchanged.

Why a config, and not plain cflags on the target:

- GN adds and removes whole configs, not individual flags, so the flags
  have to live in their own config to be removable.
- GN emits a target's own cflags before its configs' cflags, so a plain
  -mno-avx2 on the target lands before the global -mavx2 and is
  overridden. Subtracting the config is the only thing that drops them.
